+++ component/processor.go
@@ -87,14 +87,22 @@ type ProcessorFactory interface {
 	// CreateTraceProcessor creates a trace processor based on this config.
 	// If the processor type does not support tracing or if the config is not valid
 	// error will be returned instead.
-	CreateTraceProcessor(ctx context.Context, params ProcessorCreateParams,
-		nextConsumer consumer.TraceConsumer, cfg configmodels.Processor) (TraceProcessor, error)
+	CreateTraceProcessor(
+		ctx context.Context,
+		params ProcessorCreateParams,
+		cfg configmodels.Processor,
+		nextConsumer consumer.TraceConsumer,
+	) (TraceProcessor, error)
 
 	// CreateMetricsProcessor creates a metrics processor based on this config.
 	// If the processor type does not support metrics or if the config is not valid
 	// error will be returned instead.
-	CreateMetricsProcessor(ctx context.Context, params ProcessorCreateParams,
-		nextConsumer consumer.MetricsConsumer, cfg configmodels.Processor) (MetricsProcessor, error)
+	CreateMetricsProcessor(
+		ctx context.Context,
+		params ProcessorCreateParams,
+		cfg configmodels.Processor,
+		nextConsumer consumer.MetricsConsumer,
+	) (MetricsProcessor, error)
 
 	// CreateLogsProcessor creates a processor based on the config.
 	// If the processor type does not support logs or if the config is not valid